+++ src/main/java/zhang/lao/console/controller/nav/PageTurnController.java
@@ -16,9 +16,9 @@ public class PageTurnController {
     @Resource(name = "secoundSkin")
     private SkinNav nav;
     @RequestMapping("/console/nav/trun/first")
-    public String getNavFirstReturn(int nav_id, HttpServletRequest request) {
+    public String getNavFirstReturn(Long nav_id, HttpServletRequest request) {
         HttpSession httpSession =  request.getSession();
-        httpSession.setAttribute("nav", nav.gethtml(nav_id,Integer.parseInt(httpSession.getAttribute("user_id").toString()),request.getContextPath()));
+        httpSession.setAttribute("nav", nav.gethtml(nav_id,Long.parseLong(httpSession.getAttribute("user_id").toString()),request.getContextPath()));
         return "console/skins/skin_2/index";
     }
 }
