+++ src/main/java/zhang/lao/console/skin/SecoundSkin.java
@@ -36,7 +36,7 @@ public class SecoundSkin implements SkinNav{
 	/**
 	 * 获取页面导航根据系统id
 	 */
-	public String gethtml(Integer sys_id,Integer user_id,String ctxPath) {
+	public String gethtml(Long sys_id,Long user_id,String ctxPath) {
 		StringBuffer buffer=new StringBuffer();
 		SysNavExample sysNavExample = new SysNavExample();
 		sysNavExample.createCriteria().andLevelsEqualTo((short) 2).andPidEqualTo(sys_id).andStatusEqualTo((short) 1);
