+++ src/main/java/zhang/lao/mybatis/auto/dao/SysUserMapper.java
@@ -10,7 +10,7 @@ public interface SysUserMapper {
 
     int deleteByExample(SysUserExample example);
 
-    int deleteByPrimaryKey(Integer suId);
+    int deleteByPrimaryKey(Long suId);
 
     int insert(SysUser record);
 
@@ -18,7 +18,7 @@ public interface SysUserMapper {
 
     List<SysUser> selectByExample(SysUserExample example);
 
-    SysUser selectByPrimaryKey(Integer suId);
+    SysUser selectByPrimaryKey(Long suId);
 
     int updateByExampleSelective(@Param("record") SysUser record, @Param("example") SysUserExample example);
